requirements.txt
PyMuPDF>=1.24
Pillow>=10.0
tqdm>=4.0